backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Data Processing
pandas==2.1.3
numpy==1.26.2

# HTTP Requests & Scraping
requests==2.31.0
aiohttp==3.9.1
beautifulsoup4==4.12.2

# Environment & Utils
python-dotenv==1.0.0
schedule==1.2.1
tabulate==0.9.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Logging & Monitoring
logging==0.4.9.6

# FastAPI and Chart
jinja2==3.1.2
aiofiles==23.2.1